AG-10859 - Extract common financial charts logic into base class
