+++ packages/ag-charts-enterprise/src/series/candlestick/candlestickTypes.ts
@@ -1,8 +1,17 @@
-import type { AgBarSeriesStyle, LineDashOptions, StrokeOptions, _ModuleSupport } from 'ag-charts-community';
+import type {
+    AgBarSeriesStyle,
+    AgCandlestickSeriesItemType,
+    LineDashOptions,
+    StrokeOptions,
+    _ModuleSupport,
+} from 'ag-charts-community';
 
-export interface CandlestickNodeDatum
-    extends Readonly<Required<AgBarSeriesStyle>>,
-        Omit<_ModuleSupport.CartesianSeriesNodeDatum, 'yKey' | 'yValue'> {
+export interface CandlestickNodeDatum extends CandlestickNodeBaseDatum, Readonly<AgBarSeriesStyle> {
+    readonly wick?: StrokeOptions & LineDashOptions;
+}
+
+export interface CandlestickNodeBaseDatum extends Omit<_ModuleSupport.CartesianSeriesNodeDatum, 'yKey' | 'yValue'> {
+    readonly itemId: AgCandlestickSeriesItemType;
     readonly bandwidth: number;
     readonly openKey?: string;
     readonly closeKey?: string;
@@ -21,6 +30,4 @@ export interface CandlestickNodeDatum
         readonly highValue: number;
         readonly lowValue: number;
     };
-
-    readonly wick: Readonly<StrokeOptions & LineDashOptions>;
 }
